epidermis_segmentation/requirements.txt
# Core Dependencies
python>=3.8
torch>=2.0.0
torchvision>=0.15.0

# Segmentation Models
segmentation-models-pytorch>=0.3.3
albumentations>=1.3.0

# WSI Processing
openslide-python>=1.3.0
opencv-python>=4.8.0
scikit-image>=0.21.0
Pillow>=10.0.0

# Experiment Tracking
wandb>=0.15.0

# Metrics & Visualization
torchmetrics>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
pyyaml>=6.0
h5py>=3.9.0

# Additional Tools
ipython>=8.0.0
jupyter>=1.0.0
tensorboard>=2.13.0